BUG-ID: CS-48281 NPE when starting up a VM which was stopped and host was disabled after stop
Summary: In case of cold volume migration across clusters, If there are no Up and Enabled hosts available in source cluster, MigrateCommand would be allowed to use any Up and Diabled host as well. If there are no host Up in source cluster, in case of cold volume migration across clusters, the MigrateCommand would be failed.

Reviewed-By: Anshul